Fix parser error handling.
Previously there was a syntax error in builtins.swift that was causing annoying error messages in all of the tests, but wasn't causing compilation to fail.

This was because antlr was recovering from parse errors automatically, and we
weren't actually detecting it.  Now we detect all syntax errors, even those
which are automatically fixed by antlr, and abort compilation.

git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/trunk@2782 dc4e9af1-7f46-4ead-bba6-71afc04862de