Add in preliminary support for RNGs for scicolsim development.
There are currently two random functions: random(), which returns a float in the range [0.0, 1.0) and randint(a, b), which returns an integer in the range [a, b)

The randomness won't be very high quality as we're not using a proper PRNG or a good seed, but all the compiler infrastructure is there.

It will switch to running random or randint locally (instead of using the heavyweight async version) wherever possible.

In the course I found a couple of bugs in the optimizer where it missed some opportunities to fold constants

git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/trunk@2929 dc4e9af1-7f46-4ead-bba6-71afc04862de